+++ internal/testenv/docdrift_packages_test.go
@@ -45,6 +45,7 @@ var canonicalDriftTests = []string{
 	"TestDocTLSWarnsTheImageHealthcheckIsPlainHTTP",
 	"TestDocWindowsAbortIsDescribedAsFixed",
 	"TestDocLaunchdJobHasEveryFileItNames",
+	"TestDocServiceInstallStepsMatchTheUnitHeader",
 }
 
 func TestTheDocsOnlyPathStillRunsTheDriftGuards(t *testing.T) {